==> 03_classify-reactions.ipynb <==
import pandas

import automol
import chemkin_io
import helpers
import ioformat

# 1. read in the species lists
C67_SPC_PATH = 'mechanisms/01_raw_nuig-c6-7.csv'
PYR_SPC_PATH = 'mechanisms/01_raw_nuig-pyro.csv'

C67_SPC_DF = pandas.read_csv(C67_SPC_PATH, quotechar="'")
PYR_SPC_DF = pandas.read_csv(PYR_SPC_PATH, quotechar="'")

# 2. read in the reaction lists
C67_RXN_PATH = 'mechanisms/01_raw_nuig-c6-7.txt'
PYR_RXN_PATH = 'mechanisms/01_raw_nuig-pyro.txt'

C67_MECH_STR = ioformat.pathtools.read_file('.', C67_RXN_PATH)
PYR_MECH_STR = ioformat.pathtools.read_file('.', PYR_RXN_PATH)

C67_RXN_DCT = chemkin_io.parser.mechanism.reactions(C67_MECH_STR)
PYR_RXN_DCT = chemkin_io.parser.mechanism.reactions(PYR_MECH_STR)

def hash_reaction_from_chi(rct_chis, prd_chis):
    """ create a hash for a reaction from its ChIs
    """
    rct_key = automol.inchi_key.first_hash(
        automol.chi.inchi_key(automol.chi.join(rct_chis)))
    prd_key = automol.inchi_key.first_hash(
        automol.chi.inchi_key(automol.chi.join(prd_chis)))
    return f'{rct_key}_{prd_key}'


def classify_reactions(rxn_dct, spc_df):
    """ classifies reactions
    """
    row_dcts = []
    for rxn, params in rxn_dct.items():
        name = chemkin_io.writer.format_rxn_name(rxn)
        print(f'Classifying {name}...')
        
        row_dct = {
            'name': name,
            'orig-name': name,
            'params': params,
            'succeeded': False,
            'class': None,
            'path': None,
        }

        # First, classify the reaction from InChIs
        rct_names, prd_names, third_body = rxn
        rct_chis = [helpers.lookup_value(spc_df, 'inchi', {'name': n}) for n in rct_names]
        prd_chis = [helpers.lookup_value(spc_df, 'inchi', {'name': n}) for n in prd_names]

        rxn_objs = []
        fail_message = ''
        try:
            rxn_objs = automol.reac.from_chis(rct_chis, prd_chis, stereo=False)
        except Exception as e:
            fail_message = f' - Failed with error {e}'

        if not rxn_objs:
            print(' - Failed')
            print(fail_message if fail_message else ' - Unclassified')

        for num, rxn_obj in enumerate(rxn_objs):
            name = chemkin_io.writer.format_rxn_name(
                (rct_names, prd_names, third_body))

            ts_chi = automol.reac.ts_amchi(rxn_obj)

            print(' - Succeeded')
            rxn_hash = hash_reaction_from_chi(rct_chis, prd_chis)
            rxn_obj_str = automol.reac.string(rxn_obj)
            rxn_obj_path = f'reactions/graphs/{rxn_hash}_{num}.txt'
            ioformat.pathtools.write_file(rxn_obj_str, '.', rxn_obj_path)
            row_dct.update({
                'name': name,
                'succeeded': True,
                'class': rxn_obj.class_,
                'path': rxn_obj_path,
                'ts_amchi': ts_chi,
            })

        row_dcts.append(row_dct)
        
    rxn_df = pandas.DataFrame.from_records(row_dcts)
    return rxn_df

# 3. classify reactions, extract successes/failures, and write them to files
print("Classifying reactions for for NUIG-C6-7")
C67_RXN_DF = classify_reactions(C67_RXN_DCT, C67_SPC_DF)
SUCC_C67_RXN_PATH = 'reactions/03_reactions-succ_nuig-c6-7.csv'
FAIL_C67_RXN_PATH = 'reactions/03_reactions-fail_nuig-c6-7.csv'

SUCC_C67_RXN_DF = C67_RXN_DF[C67_RXN_DF['succeeded']]
FAIL_C67_RXN_DF = C67_RXN_DF[~C67_RXN_DF['succeeded']]

SUCC_C67_RXN_DF.to_csv(SUCC_C67_RXN_PATH, quotechar="'")
FAIL_C67_RXN_DF.to_csv(FAIL_C67_RXN_PATH, quotechar="'")

print("Classifying reactions for for NUIG-Pyro")
PYR_RXN_DF = classify_reactions(PYR_RXN_DCT, PYR_SPC_DF)

SUCC_PYR_RXN_PATH = 'reactions/03_reactions-succ_nuig-pyro.csv'
FAIL_PYR_RXN_PATH = 'reactions/03_reactions-fail_nuig-pyro.csv'

SUCC_PYR_RXN_DF = PYR_RXN_DF[PYR_RXN_DF['succeeded']]
FAIL_PYR_RXN_DF = PYR_RXN_DF[~PYR_RXN_DF['succeeded']]

SUCC_PYR_RXN_DF.to_csv(SUCC_PYR_RXN_PATH, quotechar="'")
FAIL_PYR_RXN_DF.to_csv(FAIL_PYR_RXN_PATH, quotechar="'")

# 4. visualize reactions
def visualize_reactions(rxn_df, spc_df):
    for num, row in rxn_df.iterrows():
        print(f'Row {num}')
        name = row['name']
        print(f'{name}')
        class_ = row['class']
        if class_ is not None:
            print(f' - Class: {class_}')
        rct_names, prd_names, _ = chemkin_io.parser.reaction.get_rxn_name(name)
        rct_smis = [helpers.lookup_value(spc_df, 'smiles', {'name': n}) for n in rct_names]
        prd_smis = [helpers.lookup_value(spc_df, 'smiles', {'name': n}) for n in prd_names]
        rxn_smi = automol.smiles.reaction(rct_smis, prd_smis)
        print(f' - SMILES: {rxn_smi}')
        ts_chi = row['ts_amchi']
        print(f' - TS AMChI: {ts_chi}')
        rct_chis = [helpers.lookup_value(spc_df, 'inchi', {'name': n}) for n in rct_names]
        prd_chis = [helpers.lookup_value(spc_df, 'inchi', {'name': n}) for n in prd_names]
        automol.chi.display_reaction(rct_chis, prd_chis)

print('Successfully classified NUIG-C6-7 reactions:')
print("WARNING: RDKit fails to properly visualize some radicals")
visualize_reactions(SUCC_C67_RXN_DF, C67_SPC_DF)

print('Successfully classified NUIG-Pyro reactions:')
print("WARNING: RDKit fails to properly visualize some radicals")
visualize_reactions(SUCC_PYR_RXN_DF, PYR_SPC_DF)

print('Failed (unclassified) NUIG-C6-7 reactions:')
print("WARNING: RDKit fails to properly visualize some radicals")
visualize_reactions(FAIL_C67_RXN_DF, C67_SPC_DF)

print('Failed (unclassified) NUIG-Pyro reactions:')
print("WARNING: RDKit fails to properly visualize some radicals")
visualize_reactions(FAIL_PYR_RXN_DF, PYR_SPC_DF)
