==> 05_generate-original-and-fully-expanded-submechanisms.ipynb <==
import pandas

import helpers
import ioformat

# 1. read in the original and fully expanded species and reaction lists
# (including only successfully classified reactions, for comparison)
ORIG_C67_SPC_PATH = 'data/mechanisms/01_raw_nuig-c6-7.csv'
ORIG_C67_RXN_PATH = 'data/reactions/03_reactions-succ_nuig-c6-7.csv'
ORIG_PYR_SPC_PATH = 'data/mechanisms/01_raw_nuig-pyro.csv'
ORIG_PYR_RXN_PATH = 'data/reactions/03_reactions-succ_nuig-pyro.csv'
FULL_C67_SPC_PATH = 'data/species/02_species-fu-exp_nuig-c6-7.csv'
FULL_C67_RXN_PATH = 'data/reactions/04_reactions-fu-exp_nuig-c6-7.csv'
FULL_PYR_SPC_PATH = 'data/species/02_species-fu-exp_nuig-pyro.csv'
FULL_PYR_RXN_PATH = 'data/reactions/04_reactions-fu-exp_nuig-pyro.csv'

ORIG_C67_SPC_DF = pandas.read_csv(ORIG_C67_SPC_PATH, quotechar="'")
ORIG_C67_RXN_DF = pandas.read_csv(ORIG_C67_RXN_PATH, quotechar="'")
ORIG_PYR_SPC_DF = pandas.read_csv(ORIG_PYR_SPC_PATH, quotechar="'")
ORIG_PYR_RXN_DF = pandas.read_csv(ORIG_PYR_RXN_PATH, quotechar="'")
FULL_C67_SPC_DF = pandas.read_csv(FULL_C67_SPC_PATH, quotechar="'")
FULL_C67_RXN_DF = pandas.read_csv(FULL_C67_RXN_PATH, quotechar="'")
FULL_PYR_SPC_DF = pandas.read_csv(FULL_PYR_SPC_PATH, quotechar="'")
FULL_PYR_RXN_DF = pandas.read_csv(FULL_PYR_RXN_PATH, quotechar="'")

# 2. generate reaction parameters dictionaries
ORIG_C67_RXN_DCT = helpers.mechanism_dict(ORIG_C67_RXN_DF)
ORIG_PYR_RXN_DCT = helpers.mechanism_dict(ORIG_PYR_RXN_DF)
FULL_C67_RXN_DCT = helpers.mechanism_dict(FULL_C67_RXN_DF)
FULL_PYR_RXN_DCT = helpers.mechanism_dict(FULL_PYR_RXN_DF)

# 3. restrict the species dataframes based on the reaction lists
ORIG_C67_SPC_DF = helpers.species_for_mechanism(ORIG_C67_SPC_DF, ORIG_C67_RXN_DCT)
ORIG_PYR_SPC_DF = helpers.species_for_mechanism(ORIG_PYR_SPC_DF, ORIG_PYR_RXN_DCT)
FULL_C67_SPC_DF = helpers.species_for_mechanism(FULL_C67_SPC_DF, FULL_C67_RXN_DCT)
FULL_PYR_SPC_DF = helpers.species_for_mechanism(FULL_PYR_SPC_DF, FULL_PYR_RXN_DCT)

# 4. write the original and fully expanded submechanisms
ORIG_C67_SPC_PATH = 'data/mechanisms/05_orig_nuig-c6-7.csv'
ORIG_C67_RXN_PATH = 'data/mechanisms/05_orig_nuig-c6-7.txt'
ORIG_PYR_SPC_PATH = 'data/mechanisms/05_orig_nuig-pyro.csv'
ORIG_PYR_RXN_PATH = 'data/mechanisms/05_orig_nuig-pyro.txt'
FULL_C67_SPC_PATH = 'data/mechanisms/05_fu-exp_nuig-c6-7.csv'
FULL_C67_RXN_PATH = 'data/mechanisms/05_fu-exp_nuig-c6-7.txt'
FULL_PYR_SPC_PATH = 'data/mechanisms/05_fu-exp_nuig-pyro.csv'
FULL_PYR_RXN_PATH = 'data/mechanisms/05_fu-exp_nuig-pyro.txt'

# a. write the species files
ORIG_C67_SPC_DF.to_csv(ORIG_C67_SPC_PATH, quotechar="'", index=False)
ORIG_PYR_SPC_DF.to_csv(ORIG_PYR_SPC_PATH, quotechar="'", index=False)
FULL_C67_SPC_DF.to_csv(FULL_C67_SPC_PATH, quotechar="'", index=False)
FULL_PYR_SPC_DF.to_csv(FULL_PYR_SPC_PATH, quotechar="'", index=False)

# b. write the reaction/mechanism files
ORIG_C67_MECH_STR = helpers.mechanism_string(ORIG_C67_RXN_DCT, ORIG_C67_SPC_DF)
ORIG_PYR_MECH_STR = helpers.mechanism_string(ORIG_PYR_RXN_DCT, ORIG_PYR_SPC_DF)
FULL_C67_MECH_STR = helpers.mechanism_string(FULL_C67_RXN_DCT, FULL_C67_SPC_DF)
FULL_PYR_MECH_STR = helpers.mechanism_string(FULL_PYR_RXN_DCT, FULL_PYR_SPC_DF)

ioformat.pathtools.write_file(ORIG_C67_MECH_STR, '.', ORIG_C67_RXN_PATH)
ioformat.pathtools.write_file(ORIG_PYR_MECH_STR, '.', ORIG_PYR_RXN_PATH)
ioformat.pathtools.write_file(FULL_C67_MECH_STR, '.', FULL_C67_RXN_PATH)
ioformat.pathtools.write_file(FULL_PYR_MECH_STR, '.', FULL_PYR_RXN_PATH)
